Fixed a bug in histogram2D_10bits

+++ Histograms/src/Histograms_py.tpp
@@ -1200,6 +1200,7 @@ np_uint32 digitizer_histogram2D_10bits_py(
 				py::array_t<DataType,py::array::c_style> data_x_in,
 				py::array_t<DataType,py::array::c_style> data_y_in)
 {
+	
 	#ifdef _WIN32_WINNT
 		uint64_t nbgroups = GetActiveProcessorGroupCount();
 		uint64_t N_t = std::min((uint64_t) 64, omp_get_max_threads()*nbgroups);
@@ -1209,7 +1210,7 @@ np_uint32 digitizer_histogram2D_10bits_py(
 
 	uint16_t* hist = (uint16_t*) malloc(N_t*sizeof(uint16_t)*(1<<20));
 	uint32_t* hist_out = (uint32_t*) malloc(sizeof(uint32_t)*(1<<20));
-	std::memset(hist,0,N_t*sizeof(uint32_t)*(1<<20));
+	std::memset(hist,0,N_t*sizeof(uint16_t)*(1<<20));
 	std::memset(hist_out,0,sizeof(uint32_t)*(1<<20));
 
 	DataType* data_x = (DataType*) data_x_in.request().ptr;
